Renamed classes with clean code directives
