+++ src/Main.java
@@ -1,5 +1,5 @@
-import application.model.Body;
-import application.model.BodyHistory;
+import application.model.body.Body;
+import application.model.body.BodyHistory;
 
 /**
  * Created by Michał on 17.03.2016.
